style: rename modules dir to core
