+++ wrapper/commands/bundle.py
@@ -7,7 +7,7 @@ from pydantic import BaseModel
 from wrapper.tools import cleaning as cm, commands as ccmd, fileoperations as fo
 from wrapper.utils import messages as msg
 from wrapper.configs import DirectoryConfig as dcfg
-from wrapper.modules import KernelBuilder, AssetsCollector
+from wrapper.core import KernelBuilder, AssetsCollector
 from wrapper.interfaces import IBundleCommand
