+++ wrapper/core/assets_collector.py
